Parse DataType from json string to object without hinting of type
So, the problem is that when a DataType is serialized into json, the
data structure that is the result has no type-information other than
a property type="ImplementationType" in the json.
To be able to read json into a Java Object we do need to tell the
objectReader what we want in return. Either you parse the json into
a Map and check the magic property, or you store the type along side the
json and use that to tell the objectMapper what type you want.

This change annotates the DataType interface with information
on what property should be checked for type information and what
implementations that are in existence. This gives us the ability to
read a json text into a DataType without knowing the type it should
result in.