DataType for accessing shared document in share-documents-request
We want to register when and by who a shared document has been accessed.
These events must contain all necessary information regarding the shared
document, because metadata about the shared document itself is not known
once the sharing has stopped. This enables both sender and receiver to
see a history of access even after the sharing is stopped.

Additionally, the SharingStopped type is extended with optional fields
explaining which party stopped the sharing. This is especially important
for senders, who may be confused by the state of their sharing request
in the event that the receiver stops the sharing before the sender has
had the chance to access the shared documen.